Add audio skip voting
This commit implements voting to skip an audio. The /skip command
now adds your vote to a total, and if the total reaches the
threshold (set by the first vote), then the audio is skipped.

The voting status is also shown on the playing audio message.

The skip button has also been removed from the playing audio message,
along with the toggle loop button.